requirements.txt
Django==1.9.7
django-filter==0.13.0
djangorestframework==3.3.3
MySQL-python==1.2.5